Force to downgrade some extra PVIVs
If an IV is upgraded to PVIV at compile time
we are not going to store the PVIV correctly.

The PV slot can be empty and we are going to try
to save a CowPV and use incorrect flags.
The flag 'SVf_PROTECT' is going to be lost among others.

The best alternative is to force a downgrade in such
scenario.